+++ tasks/src/robocorp/tasks/_exceptions.py
@@ -6,3 +6,7 @@ class RobocorpTasksCollectError(RobocorpTasksError):
     """
     Exception given if there was some issue collecting tasks.
     """
+
+
+class InvalidArgumentsError(RobocorpTasksError):
+    pass